DAO refactoring DAO unit tests JavaDocs

+++ EPAMProjects/HotelManagement/src/main/java/ua/cc/spon/db/entity/Locale.java
@@ -4,12 +4,32 @@ import lombok.Data;
 
 import java.io.Serializable;
 
+/**
+ * Class describes Locale entity.
+ *
+ * @author Sergiy Ponomarenko
+ */
 @Data
 public class Locale implements Entity, Serializable {
 
-    private long id;
+    private int id;
+
+    /**
+     * Short Locale name.
+     * Ex.: uk, en, de
+     */
     private String name;
+
+    /**
+     * Country icon path.
+     * Format: flag-country-ua
+     */
     private String iconPath;
+
+    /**
+     * Full Locale name in native language.
+     * Ex.: English, Українська, Deutsch.
+     */
     private String fullName;
 
 }
