+++ EPAMProjects/HotelManagement/src/main/java/ua/cc/spon/db/entity/Request.java
@@ -6,10 +6,17 @@ import java.io.Serializable;
 import java.time.LocalDate;
 import java.util.List;
 
+/**
+ * Class describes Request entity.
+ * Request is application from User to Manager to choose the most favorite Room
+ * based on request parameters.
+ *
+ * @author Sergiy Ponomarenko
+ */
 @Data
 public class Request implements Entity, Serializable {
 
-    private long id;
+    private int id;
     private LocalDate checkinDate;
     private LocalDate checkoutDate;
     private int persons;
