+++ EPAMProjects/HotelManagement/src/main/java/ua/cc/spon/db/entity/RoomCategory.java
@@ -5,10 +5,14 @@ import lombok.Data;
 import java.io.Serializable;
 import java.sql.Timestamp;
 
+/**
+ *
+ * @author Sergiy Ponomarenko
+ */
 @Data
 public class RoomCategory implements Entity, Serializable {
 
-    private long id;
+    private int id;
     private String name;
     private String description;
     private Timestamp creationDate;
